Update README.md
"Catch the Mole" is an entertaining and interactive web-based game that challenges players' reflexes and agility. This GitHub repository houses the complete source code for the game. 

In this game, players are tasked with catching mischievous moles while avoiding dangerous plants. The game features a 3x3 grid of tiles, each randomly populated with either a mole or a plant at specific intervals. Players must click on the mole as quickly as possible to earn points, but they must avoid clicking on the plant to prevent their score from decreasing.

To add an extra layer of excitement, a countdown timer is included, adding a sense of urgency to the gameplay. Players have a limited amount of time to catch as many moles as they can and achieve the highest score possible before the timer reaches zero.
Developers and enthusiasts can use this code as a starting point for creating their own web-based games or simply as an educational resource to learn about HTML, CSS, and JavaScript game development.
Experience the thrill of "Catch the Mole" by exploring this repository and enjoy the fun of hunting moles while avoiding perilous plants!